chore: Move to utils
