+++ src/genai/streamlit_pages/utils.py
@@ -13,3 +13,13 @@ def delete_messages_state(key: Optional[str] = None) -> None:
         del st.session_state["location"]
     except KeyError:
         pass
+
+
+def reset_state() -> None:
+    """Delete the message placeholder state."""
+    keys = ["areas_of_learning_text", "areas_of_learning", "n_results", "location", "messages"]
+    for key in keys:
+        try:
+            del st.session_state[key]
+        except KeyError:
+            pass
